requirements.txt
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
et_xmlfile==2.0.0
idna==3.10
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
six==1.17.0
soupsieve==2.7
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
